requirements.txt
diffusers[torch]
torch
gradio
pyyaml
Pillow
